backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
anthropic==0.30.1
pypdf==4.3.1
python-docx==1.1.2
python-dotenv==1.0.1
werkzeug==3.0.3
